First success for indexing titles of opened-during-the-lifetime-of-the-install pages.
There are a few choices around when to invalidate the user activity.
 * when the tab closes
 * after each location change and when the tab closes

Currently maintaining a model of one user activity per tab; though the call to `activity.becomeCurrent()` should be done when the tab becomes visible.

I would quite like to beef up the BrowserHelper to be a one-to-many delegate for Browser.

We probably shouldn't be indexing all pages (detect form submission? piggy back on the readability check? index only reader mode text?).